+++ crates/polars-arrow/src/legacy/array/mod.rs
@@ -1,4 +1,7 @@
-use crate::array::{Array, BinaryArray, BooleanArray, ListArray, PrimitiveArray, Utf8Array};
+use crate::array::{
+    new_null_array, Array, BinaryArray, BooleanArray, FixedSizeListArray, ListArray,
+    PrimitiveArray, StructArray, Utf8Array,
+};
 use crate::bitmap::MutableBitmap;
 use crate::datatypes::ArrowDataType;
 use crate::legacy::utils::CustomIterTools;
@@ -15,6 +18,8 @@ pub mod slice;
 pub mod utf8;
 
 pub use slice::*;
+
+use crate::legacy::prelude::LargeListArray;
 
 macro_rules! iter_to_values {
     ($iterator:expr, $validity:expr, $offsets:expr, $length_so_far:expr) => {{
@@ -206,3 +211,53 @@ pub trait PolarsArray: Array {
 }
 
 impl<A: Array + ?Sized> PolarsArray for A {}
+
+fn is_nested_null(data_type: &ArrowDataType) -> bool {
+    match data_type {
+        ArrowDataType::Null => true,
+        ArrowDataType::LargeList(field) => is_nested_null(field.data_type()),
+        ArrowDataType::FixedSizeList(field, _) => is_nested_null(field.data_type()),
+        ArrowDataType::Struct(fields) => {
+            fields.iter().all(|field| is_nested_null(field.data_type()))
+        },
+        _ => false,
+    }
+}
+
+/// Cast null arrays to inner type and ensure that all offsets remain correct
+pub fn convert_inner_type(array: &dyn Array, dtype: &ArrowDataType) -> Box<dyn Array> {
+    match dtype {
+        ArrowDataType::LargeList(field) => {
+            let array = array.as_any().downcast_ref::<LargeListArray>().unwrap();
+            let inner = array.values();
+            let new_values = convert_inner_type(inner.as_ref(), field.data_type());
+            let dtype = LargeListArray::default_datatype(new_values.data_type().clone());
+            LargeListArray::new(
+                dtype,
+                array.offsets().clone(),
+                new_values,
+                array.validity().cloned(),
+            )
+            .boxed()
+        },
+        ArrowDataType::FixedSizeList(field, width) => {
+            let array = array.as_any().downcast_ref::<FixedSizeListArray>().unwrap();
+            let inner = array.values();
+            let new_values = convert_inner_type(inner.as_ref(), field.data_type());
+            let dtype =
+                FixedSizeListArray::default_datatype(new_values.data_type().clone(), *width);
+            FixedSizeListArray::new(dtype, new_values, array.validity().cloned()).boxed()
+        },
+        ArrowDataType::Struct(fields) => {
+            let array = array.as_any().downcast_ref::<StructArray>().unwrap();
+            let inner = array.values();
+            let new_values = inner
+                .iter()
+                .zip(fields)
+                .map(|(arr, field)| convert_inner_type(arr.as_ref(), field.data_type()))
+                .collect::<Vec<_>>();
+            StructArray::new(dtype.clone(), new_values, array.validity().cloned()).boxed()
+        },
+        _ => new_null_array(dtype.clone(), array.len()),
+    }
+}
